% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trouees.R
\name{data_trouees}
\alias{data_trouees}
\title{Cartographie des trouées}
\usage{
data_trouees(mnh, ext, h_min = 3, min_size = 300, cvt_max = 10)
}
\arguments{
\item{mnh}{spatraster du MNH}

\item{ext}{sf polygone de la zone à cartographier}

\item{h_min}{hauteur à partir du sol où évaluer le couvert}

\item{min_size}{taille minimale des trouées à cartographier en m2}

\item{cvt_max}{couvert maximum (%) à l'échelle de l'are des peuplements ouverts}
}
\value{
spatraster
}
\description{
Cartographie des trouées
}
