% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ref_crowns_params_placette.R
\name{md.ref_crowns_params_placette}
\alias{md.ref_crowns_params_placette}
\title{Fonction interne de md_ref_crowns_params}
\usage{
md.ref_crowns_params_placette(
  i,
  spec_date,
  aci,
  arbres_ttl,
  fake_proj,
  grille_lidarhd,
  topo,
  dendro,
  spot,
  insol,
  mnh,
  sentinel,
  copernicus,
  dest_pp,
  rsp_coper = FALSE
)
}
\arguments{
\item{i}{ligne de aci}

\item{spec_date}{date}

\item{aci}{table des placettes}

\item{arbres_ttl}{table des arbres}

\item{fake_proj}{nom projet}
}
\value{
sf
}
\description{
Fonction interne de md_ref_crowns_params
}
