% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_calibr_fertilite.R
\name{md_calibr_fertilite}
\alias{md_calibr_fertilite}
\title{Récolte de données de calibration pour référence de fertilité}
\usage{
md_calibr_fertilite(
  path_calibr = "/var/partage/oiseau/modeles/calib_fertilite.rds",
  add = TRUE,
  spat_filter = "CCOD_FRT",
  by = "all",
  plot = FALSE
)
}
\arguments{
\item{path_calibr}{chemin du fichier de calibration}
}
\value{
test des modèles d'accroissement de référence
}
\description{
A executer au sein de projets utilisés comme référence
}
