% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_crowns_calibration.R
\name{md_crowns_calibration}
\alias{md_crowns_calibration}
\title{Création d'un modèle calibré par des points d'observation}
\usage{
md_crowns_calibration(
  path_points,
  y_name,
  dest_name = y_name,
  title = "calibration",
  date = "last",
  select_vars = TRUE,
  exclude = c("dend"),
  exclude_seuil = 0.05,
  path_crowns = data_conf("path_crowns_ts"),
  path_spot = data_conf("path_spot_ts"),
  dest_dos = dc("dos_modeles")
)
}
\arguments{
\item{path_points}{chemin vers le shapefile de points de calibration}

\item{y_name}{nom du champ de la variable à prédire}

\item{dest_name}{nom du raster en sortie}

\item{title}{Nom donné au modèle}

\item{date}{date du MNH à utiliser}

\item{select_vars}{TRUE pour Utiliser une sélection automatique des variable}

\item{exclude}{exclure des prédicteurs les familles dend(rométriques) topo(graphiques) spot insolation mnh area (surface)}

\item{exclude_seuil}{taux de perte max de placettes tolérée en raison de données absentes avant exclusion de la variable}

\item{path_crowns}{chemin du raster des couronnes}

\item{path_spot}{chemin du rster des images spot}

\item{dest_dos}{chemin du dossier des modèles}
}
\value{
modèle RF
}
\description{
Création d'un modèle calibré par des points d'observation
}
