% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_crowns_classify.R
\name{md_crowns_classify}
\alias{md_crowns_classify}
\title{Classification automatique des couronnes}
\usage{
md_crowns_classify(
  ncat = NULL,
  rm_size = FALSE,
  dest_file = "class_crowns.tif",
  path_crowns = data_conf("path_crowns_ts"),
  dos_metrics = data_conf("dos_modeles"),
  file_metrics = "crowns_metrics_.*.rds",
  dest_dos = dc("dos_spot"),
  plot = FALSE,
  methode = "kmeans"
)
}
\arguments{
\item{ncat}{nb de types à créer}

\item{rm_size}{TRUE pour retirer les dimensions (h et surface) des facteurs explicatifs}

\item{dest_file}{nom du fichier à créer (avec extention .tif)}

\item{path_crowns}{chemin du raster des couronnes}

\item{dest_dos}{dossier d'écriture}

\item{path_spot}{chemin du rster des images spot}
}
\value{
spatRaster de catégories
}
\description{
Classification automatique des couronnes
}
