% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_downscale.R
\name{md_downscale}
\alias{md_downscale}
\title{Réechantillonage d'un raster 10m à 1m par régression linéaire}
\usage{
md_downscale(y10m10, an = "last", rf = NULL)
}
\arguments{
\item{y10m10}{raster numérique de résolution 10m}

\item{an}{année spot à utiliser}
}
\value{
spatrast 1m
}
\description{
Réechantillonage d'un raster 10m à 1m par régression linéaire
}
