% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_large_echantillon.R
\name{md_large_echantillon}
\alias{md_large_echantillon}
\title{Echantillonnage de rasters à large échelle}
\usage{
md_large_echantillon(
  rs,
  ans = NULL,
  nspl = 10,
  larg_grappe = 1000,
  rsp_coper = FALSE,
  dest = dc("dos_modeles_oiseau")
)
}
\arguments{
\item{rs}{pile de raster (y) pour lesquels on cherche un échantillon de prédicteurs}

\item{nspl}{nb de grappes à utiliser}

\item{larg_grappe}{largeur de la garppe carrée}

\item{an}{an spot utilisée}
}
\value{
liste de
}
\description{
Echantillonnage de rasters à large échelle
}
