% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_metrics_selection.R
\name{md_metrics_selection}
\alias{md_metrics_selection}
\title{Sélection de variables numériques}
\usage{
md_metrics_selection(data, y, big = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{y}{champ à prédire}
}
\value{
nom des variables sélectionnées
}
\description{
Sélection de variables numériques
}
