% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_metrics_used.R
\name{md_metrics_used}
\alias{md_metrics_used}
\title{Premier filtre des namess de prédicteurs}
\usage{
md_metrics_used(
  names,
  exclu = c("mnh_aspect", "flowdir", "mnh_TPI", "spot_.*_q", "moy", "topo_.*q",
    "topo_.*sd", ".*_x_.*q", ".*_x_.*sd", "TRIrmsd", "TRIriley")
)
}
\arguments{
\item{names}{vecteur de nom de variables}

\item{exclu}{expressions exclues}
}
\value{
vecteur filtré
}
\description{
Premier filtre des namess de prédicteurs
}
