% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mlr3.R
\name{md_mlr3_compare}
\alias{md_mlr3_compare}
\title{Comparatif des méthode d'apprentissage}
\usage{
md_mlr3_compare(
  data,
  y,
  op = "classif",
  meth = c("kknn", "naive_bayes", "nnet", "glmnet", "svm", "rpart", "featureless",
    "lda", "log_reg", "multinom", "qda", "ranger", "xgboost"),
  options = list(rpart = "cp = 0.01")
)
}
\arguments{
\item{data}{table}

\item{y}{nom du champ à prédire}

\item{op}{classif, clust ou regr}

\item{meth}{liste des méthodes à tester: 2nde partie de la clé d'apprentissage. Voir lrns()}

\item{options}{liste des options. Ex: list(rpart = "cp = 0.01")}
}
\value{

}
\description{
Comparatif des méthode d'apprentissage
}
