% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mlr3.R
\name{md_mlr3_compare_as_table}
\alias{md_mlr3_compare_as_table}
\title{Présentation des données issues de md_mlr3_compare()}
\usage{
md_mlr3_compare_as_table(ls)
}
\arguments{
\item{ls}{liste issue de md_mlr3_compare()}
}
\value{
table
}
\description{
Présentation des données issues de md_mlr3_compare()
}
