% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mlr3_predicteurs.R
\name{md_mlr3_predicteurs}
\alias{md_mlr3_predicteurs}
\title{Construction de la pile des prédicteurs utilisés par un modèle MLR3}
\usage{
md_mlr3_predicteurs(
  fea,
  res = NULL,
  spot_an = NULL,
  rsp_coper = FALSE,
  dest = file.path(dc("dos_projet"), "pile_essences.tif")
)
}
\arguments{
\item{fea}{variables sélectionnée (cf md_mlr3_select)}

\item{res}{résolution des rasters à compiler}

\item{spot_an}{année de l'image spot utilisée}

\item{rsp_coper}{T/F resampler les rasters Copernicus à la résolution res}

\item{dest}{chemin d'écriture de la pile}
}
\value{
spatraster
}
\description{
Construction de la pile des prédicteurs utilisés par un modèle MLR3
}
