% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mlr3.R
\name{md_mlr3_select}
\alias{md_mlr3_select}
\title{Modèle par package mlr3: selection des variables explicatives}
\usage{
md_mlr3_select(data, m, y, op = "classif", opt = NULL, n_max = 8, test = FALSE)
}
\arguments{
\item{data}{table}

\item{m}{méthode: 2nde partie de la clé d'apprentissage. Voir lrns()}

\item{y}{nom du champ à prédire}

\item{op}{classif, clust ou regr}

\item{opt}{liste des options. Ex: list(rpart = "cp = 0.01")}

\item{n_max}{nb max à sélectionner}

\item{test}{TRUE pour md_mlr3_compare}
}
\value{
liste des variables retenues
}
\description{
Modèle par package mlr3: selection des variables explicatives
}
