% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_modele_vigueur.R
\name{md_modele_vigueur}
\alias{md_modele_vigueur}
\title{Création modèle croissance / hauteur}
\usage{
md_modele_vigueur(
  path_calibr = "/var/partage/oiseau/modeles/calib_fertilite.rds"
)
}
\arguments{
\item{path_calibr}{chemin du fichier de calibration}
}
\value{
écrit modele_vigueur.rds
}
\description{
Création modèle croissance / hauteur
}
