% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_new_cah.R
\name{md_new_classif}
\alias{md_new_classif}
\title{Classification automatique par CAH ou kmeans}
\usage{
md_new_classif(data, methode = "kmeans", nb_cat = NULL)
}
\arguments{
\item{data}{table}

\item{methode}{kmeans ou cah}

\item{nb_cat}{NULL ou nombre de classes voulu}
}
\value{
liste: cats = catégories, vars = variables représentatives
}
\description{
Classification automatique par CAH ou kmeans
}
