% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_new_model_rf.R
\name{md_new_model_rf}
\alias{md_new_model_rf}
\title{md_new_model_rf}
\usage{
md_new_model_rf(tab, col_y, pvalid = 0.2, nb_cat = NULL)
}
\arguments{
\item{tab}{table des prédicteurs}

\item{col_y}{nom de la colonne à prédire}

\item{pvalid}{taux de lignes à conserver pour la validAtion}
}
\value{
modèle RF
}
\description{
md_new_model_rf
}
