% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_predict_crowns.R
\name{md_predict_crowns}
\alias{md_predict_crowns}
\title{Prédiction appliquée sur les couronnes}
\usage{
md_predict_crowns(
  mn,
  dir,
  res = NULL,
  spot_an = NULL,
  dest = file.path(dc("dos_modeles"))
)
}
\arguments{
\item{mn}{nom du modèle: nom du fichier où le modèle est enregistré, sans extension .rds}

\item{dir}{dossier du fichier}

\item{res}{resolution des rasters d'extraction (defaut tiré du nom du fichier ...-res1-... = 1m)}

\item{spot_an}{date de l'image spot uitilisée (defaut tiré du nom du fichier ...-spot2018-... = 2018)}
}
\value{
spatraster
}
\description{
Prédiction appliquée sur les couronnes
}
