% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_predicteurs.R
\name{md_predicteurs}
\alias{md_predicteurs}
\title{Raster de prédicteurs pour spatialisation}
\usage{
md_predicteurs(
  topo = TRUE,
  dendro = TRUE,
  spot = TRUE,
  insol = TRUE,
  mnh = TRUE,
  sentinel = TRUE,
  suppress_rasters_with_na = 100,
  res = 10,
  spot_date = "last"
)
}
\arguments{
\item{suppress_rasters_with_na}{retire des prédicteurs les rasters dont le % de valeurs NA est supérieur au nombre donné (100: garde tout, 0: supprime si une valeur absente)}

\item{res}{resolution du pixel en m}
}
\value{
spatRaster
}
\description{
Raster de prédicteurs pour spatialisation
}
