% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_process_essence.R
\name{md_process_essence}
\alias{md_process_essence}
\title{Création interactive du raster des essences}
\usage{
md_process_essence(defaut_train = "crown_ess_new2_ain_res1-spot2018.rds")
}
\arguments{
\item{defaut_train}{nom}
}
\value{
rien, raster accessible avec data_crowns_essences()
}
\description{
Création interactive du raster des essences
}
