% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_process_essence.R
\name{md_process_essence_new_model}
\alias{md_process_essence_new_model}
\title{Création d'un modèle d'essences}
\usage{
md_process_essence_new_model(
  defaut_train,
  dest = dc("dos_modeles_oiseau"),
  name = NULL
)
}
\arguments{
\item{defaut_train}{nom du jeu de calibration utilisé par défaut}

\item{dest}{dossier des modèles oiseau}

\item{name}{nom à donner au modèle}
}
\value{

}
\description{
Création d'un modèle d'essences
}
