% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_process_essence.R
\name{md_process_essence_predict}
\alias{md_process_essence_predict}
\title{Prédiction de l'essence des couronnes}
\usage{
md_process_essence_predict(
  nom_modele = NULL,
  dos_modele = dc("dos_modeles_oiseau")
)
}
\arguments{
\item{nom_modele}{modèle, sans prefix mlr3ess_ et sans extention .rds}

\item{dos_modele}{dossier}
}
\value{
table, cr=id couronne, pred=prediction
}
\description{
Prédiction de l'essence des couronnes
}
