% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_process_essence.R
\name{md_process_essence_rasterize}
\alias{md_process_essence_rasterize}
\title{Raster des essences à res=1m}
\usage{
md_process_essence_rasterize(tab, dest = dc("path_essences"))
}
\arguments{
\item{tab}{cr = id crown (cf data_crowns), pred = code essence à 3 caractères}

\item{dest}{chemin d'écriture du tif}
}
\value{
écrit tif et csv des cat
}
\description{
Raster des essences à res=1m
}
