% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_process_essence.R
\name{md_process_essence_select_train}
\alias{md_process_essence_select_train}
\title{Choix des données de calibration des modèles essences}
\usage{
md_process_essence_select_train(
  path_cal = NULL,
  nom_ref = NULL,
  defaut_train = NULL
)
}
\arguments{
\item{path_cal}{NULL ou chemin des couronnes de calibration créés avec md_ref_crowns_params()}

\item{nom_ref}{nom du jeu de calibration ou NULL pour choix interactif}

\item{defaut_train}{jeu par défaut}
}
\value{
table
}
\description{
Choix des données de calibration des modèles essences
}
