% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ref_crowns_params.R
\name{md_ref_crowns_params}
\alias{md_ref_crowns_params}
\title{Métriques des houppiers des arbres des placettes permanentes}
\usage{
md_ref_crowns_params(
  nom_ref = "ess_ain",
  placettes = dbOiseau::bdloc_read_table("placettes_permanentes_placettes") \%>\%
    filter(!is.na(date) & date > as.Date("2015-01-01")),
  arbres_ttl = dbOiseau::bdloc_read_table("placettes_permanentes_arbres") \%>\%
    filter(is.na(type)),
  spec_date = 2018,
  sentinel = FALSE,
  dest_dir = dc("dos_modeles_oiseau"),
  dest_name = paste0("crown_", nom_ref, "_res1-spot", spec_date),
  user = basename(dirname(dc("dos_projet"))),
  fake_proj = "tmp_ref_crowns"
)
}
\arguments{
\item{nom_ref}{nom du jeu de données}

\item{placettes}{table des placettes permanentes, avec champ numdisp (identifiant réseau) et numplac (identifiant placette au sein du réseau)}

\item{arbres_ttl}{table des arbres avec champ de coordonnées polaire azimut (en grad) et dist (en m) et champs placette (numdisp et numpla)}

\item{spec_date}{date des images spot à utiliser}

\item{dest_dir}{dossier d'écriture}

\item{dest_name}{nom de fichier d'écriture}

\item{user}{nom utilisateur}

\item{fake_proj}{nom du projet fictif de travail à créer}
}
\value{
rien, écrit le fichier .rds des données de calibration
}
\description{
Métriques des houppiers des arbres des placettes permanentes
}
