% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_resample_copernicus.R
\name{md_resample_copernicus}
\alias{md_resample_copernicus}
\title{Modèle de rééchantillonnage des rasters Copernicus}
\usage{
md_resample_copernicus(dalle)
}
\arguments{
\item{dalle}{numéro de la dalle}
}
\value{

}
\description{
Modèle de rééchantillonnage des rasters Copernicus
}
