% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_rm_cor_var.R
\name{md_rm_corr_var}
\alias{md_rm_corr_var}
\title{Suppresion de variables hautement corrélées}
\usage{
md_rm_corr_var(df, lim = 0.9)
}
\arguments{
\item{df}{table}

\item{lim}{correlation max tolérée}
}
\value{

}
\description{
Suppresion de variables hautement corrélées
}
