% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_spatialise.R
\name{md_spatialise}
\alias{md_spatialise}
\title{Cartographie par apprentissage}
\usage{
md_spatialise(
  fac2pred,
  tab_inv = util_read_csv(dc("tab_invrel_data")),
  coords = c("Coord.X", "Coord.Y"),
  epsg = 4326,
  buff_pla = 20,
  use_topo = TRUE,
  use_dendro = TRUE,
  use_spot = TRUE,
  use_insol = TRUE,
  use_couvert = TRUE,
  suppress_rasters_with_na = 0,
  rsp_coper = FALSE,
  list_fac = FALSE
)
}
\arguments{
\item{fac2pred}{champs de la table tab_invrel_data à cartographier}

\item{buff_pla}{rayon d'extraction}

\item{use_topo}{si TRUE, utilise comme prédicteurs les rasters topographiques dérivés du MNT}

\item{use_dendro}{si TRUE, utilise comme prédicteurs les rasters dendrométriques issus des modèles lidar}

\item{use_spot}{si TRUE, utilise comme prédicteurs les indices de végétation spot}

\item{use_insol}{si TRUE, utilise comme prédicteurs les rasters d'insolation calculés par spot_insolation()}

\item{use_couvert}{si TRUE, utilise comme prédicteurs le MNH}

\item{suppress_rasters_with_na}{retire des prédicteurs les rasters dont le % de valeurs NA est supérieur au nombre donné (100: garde tout, 0: supprime si une valeur absente)}

\item{list_fac}{TRUE pour lister les facteurs disponibles}
}
\value{
spatRaster
}
\description{
Cartographie par apprentissage
}
