% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnh_select_year.R
\name{mnh_select_year}
\alias{mnh_select_year}
\title{Raster MNHJ d'une année donnée}
\usage{
mnh_select_year(
  an,
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  ifnull = c("near", "nothing")
)
}
\arguments{
\item{an}{muneric}

\item{path_mnh_ts}{chemin de la série temporelle spot de l'emprise du projet}

\item{ifnull}{si non trouvée, chercher l'année la plus proche ("near") ou rien ("nothing")}
}
\value{
spatraster
}
\description{
Raster MNHJ d'une année donnée
}
