% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spa_select_vars.R
\name{spa_select_vars}
\alias{spa_select_vars}
\title{Sélection automaatique de variables}
\usage{
spa_select_vars(x, y, family = NULL, cutoff = 0.75, nb_fact = 5)
}
\arguments{
\item{nb_fact}{}
}
\value{

}
\description{
Sélection automaatique de variables
}
