% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_clouds.R
\name{spot.clouds}
\alias{spot.clouds}
\title{Masque de nuage d'une image SPOT}
\usage{
spot.clouds(path)
}
\arguments{
\item{path}{Chemin du tif spot}
}
\value{
sf avec champ maskType (CLOUD ou SNOW) ou NULL (pas de masque dispo) ou "ko" (erreur)
}
\description{
Masque de nuage d'une image SPOT
}
