% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_an_prec.R
\name{spot_an_prec}
\alias{spot_an_prec}
\title{Année précédente disponible la plus proche d'une année donnée}
\usage{
spot_an_prec(n, an)
}
\arguments{
\item{n}{liste des noms des rasters spot}

\item{an}{année de référence}
}
\value{
integer
}
\description{
Année précédente disponible la plus proche d'une année donnée
}
