% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_apex.R
\name{spot_apex}
\alias{spot_apex}
\title{Détection des apex sur MNH}
\usage{
spot_apex(h, lim_h_rege = 3, pente = 0.07, intercept = 2)
}
\arguments{
\item{h}{mnh projeté en Lambert93 (epsg 2154)}

\item{lim_h_rege}{hauteur minimale des apex à détecter}

\item{pente}{voir lidR::locate_trees}

\item{intercept}{voir lidR::locate_trees}
}
\value{
spatialpoints
}
\description{
Détection des apex sur MNH
}
