% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_conf_ini.R
\name{spot_conf_ini}
\alias{spot_conf_ini}
\title{Configuration par défaut d'un projet pour le calcul du couvert}
\usage{
spot_conf_ini(
  .dir = mget(".dir", envir = as.environment(1), ifnotfound = list(NULL))[[1]],
  replace = FALSE
)
}
\arguments{
\item{.dir}{dossier du projet}

\item{replace}{TRUE pour réinitialiser les paramètres}
}
\value{

}
\description{
Configuration par défaut d'un projet pour le calcul du couvert
}
