% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_crowns_calibration.R
\name{spot_crowns_calibration}
\alias{spot_crowns_calibration}
\title{Création d'un modèle calibré par des points d'observation}
\usage{
spot_crowns_calibration(
  path_points,
  y_name,
  title = "calibration",
  select_vars = TRUE,
  path_crowns = oiseauData::data_conf("path_crowns_ts"),
  path_spot = oiseauData::data_conf("path_spot_ts"),
  dest_dos = dc("dos_modeles")
)
}
\arguments{
\item{path_points}{chemin vers le shapefile de points de calibration}

\item{y_name}{nom du champ de la variable à prédire}

\item{title}{Nom donné au modèle}

\item{select_vars}{TRUE pour Utiliser une sélection automatique des variable}

\item{path_crowns}{chemin du raster des couronnes}

\item{path_spot}{chemin du rster des images spot}

\item{dest_dos}{chemin du dossier des modèles}
}
\value{
modèle RF
}
\description{
Création d'un modèle calibré par des points d'observation
}
