% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_crowns_metrics.R
\name{spot_crowns_metrics}
\alias{spot_crowns_metrics}
\title{Rasters des métriques des couronnes}
\usage{
spot_crowns_metrics(
  date_mnh = util_get_date("last", "mnh"),
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  path_crowns_ts = oiseauData::data_conf("path_crowns_ts"),
  path_meta = oiseauData::data_conf("tab_crowns"),
  path_mnt = oiseauData::data_conf("path_mnt"),
  buffer = oiseauData::data_conf("buffer"),
  mask = oiseauData::data_conf("shp"),
  lim_h_rege = oiseauData::data_conf("lim_h_rege"),
  best_day = oiseauData::data_conf("spot_best_day"),
  dest_dir = dc("dos_modeles"),
  spot_date = "last",
  sentinel = TRUE
)
}
\arguments{
\item{date_mnh}{date du MNH à utiliser}

\item{crowns}{raster des couronnes issu de spot_crowns}

\item{mnh}{}
}
\value{
spatRaster
}
\description{
Rasters des métriques des couronnes
}
