% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_data.R
\name{spot_data}
\alias{spot_data}
\title{Création image spot6/7 1.5m d'une zone}
\usage{
spot_data(dos_spot, roi, destpath, buffer = 100)
}
\arguments{
\item{dos_spot}{dossier des images spot6/7}

\item{roi}{sf de la zone d'intérêt. Si NULL, renvoie seulement la liste des polygones des emprises}

\item{destpath}{chemin du dossier où écrire les images}

\item{buffer}{buffer à appliquer à roi}
}
\value{
liste des polygones es emprises des dalles spots
}
\description{
Création image spot6/7 1.5m d'une zone
}
