% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_disparitions.R
\name{spot_disparitions}
\alias{spot_disparitions}
\title{Raster des couronnes disparues entre deux images spot}
\usage{
spot_disparitions(
  date0,
  date1,
  date_mnh,
  ext = oiseauData::data_conf("shp"),
  path_spot_ts = oiseauData::data_conf("path_spot_ts"),
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  tab_mnh = oiseauData::data_conf("tab_mnh"),
  buffer = oiseauData::data_conf("buffer"),
  seuil_diff_spot = oiseauData::data_conf("seuil_diff_spot")
)
}
\arguments{
\item{date0}{date de la première image spot}

\item{date1}{date de la seconde image spot}

\item{date_mnh}{date du MNH utilisé pour la détection des couronnes}

\item{ext}{sf de la zone}

\item{path_spot_ts}{chemin du fichier .nc d'écriture de la série temporelle}

\item{path_mnh_ts}{chemin du MNH}

\item{buffer}{buffer shp}

\item{seuil_diff_spot}{seuil de différence à partir duquel la couronne est éliminée}
}
\value{
spatraster 0 = disparu, 1 = toujours présent
}
\description{
Raster des couronnes disparues entre deux images spot
}
