% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_formula.R
\name{spot_formula}
\alias{spot_formula}
\title{Formules d'indices de végétation SPOT}
\usage{
spot_formula(indice)
}
\arguments{
\item{indice}{nom de l'indice. Voir https://www.indexdatabase.de/db/s-single.php?id=173}
}
\value{
texte de la formule
}
\description{
Formules d'indices de végétation SPOT
}
