% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_get.R
\name{spot_get}
\alias{spot_get}
\title{Série temporelle des images spot}
\usage{
spot_get(paths, ext = NULL, buffer = 10, ignore = NULL, template = NULL)
}
\arguments{
\item{paths}{liste des chemins d'images spot d'une même tuile. Si non généré par spot_data(), nommée par l'année de prise de vue}

\item{ext}{sf de découpage}

\item{buffer}{buffer à appliquer à ext}

\item{ignore}{liste des dates à ignorer, au format 2022-02-25}

\item{template}{spatraster sur lequel s'aligner}
}
\value{
liste de spatRasters 4 bandes
}
\description{
Série temporelle des images spot
}
