% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_indice.R
\name{spot_indice}
\alias{spot_indice}
\title{Indice de végétation des couronnes}
\usage{
spot_indice(
  date_mnh,
  date_spot,
  shp_roi = oiseauData::data_conf("shp"),
  path_spot_ts = oiseauData::data_conf("path_spot_ts"),
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  buffer = oiseauData::data_conf("buffer"),
  indice = "ndvi"
)
}
\arguments{
\item{date_mnh}{date du MNH utilisé}

\item{date_spot}{date de l'image SPOT utilisée}

\item{shp_roi}{sf de découpage}

\item{path_spot_ts}{chemin vers le raster de série temporelle spot}

\item{path_mnh_ts}{chemin vers le raster de série temporelle du MNH (résolution 1m)}

\item{buffer}{buffer à appliquer à shp_roi}

\item{indice}{nom de l'indice utilisé, par défaut ndvi. voir spot_formula}
}
\value{
spatRaster
}
\description{
Indice de végétation des couronnes
}
