% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_perturbation.R
\name{spot_perturbation}
\alias{spot_perturbation}
\title{Test de délimitation des zones ouvertes entre 2 dates}
\usage{
spot_perturbation(
  lim,
  diff = NULL,
  ras_dif = NULL,
  liste_spots = NULL,
  an,
  sub = NULL,
  horizontal = TRUE,
  overlay = NULL,
  mask = FALSE,
  col = "blue"
)
}
\arguments{
\item{lim}{valeur limite de l'indicateur utilisé par ras_diff}

\item{diff}{liste des rasters sortis de spot_differences (ignoré si ras_dif non nul)}

\item{ras_dif}{si diff NULL, spatraster des différences entre deux images spot}

\item{liste_spots}{liste des rasters RGB Spot, nommée "spotXXXX" où XXXX est l'année de prise de vue}

\item{an}{année à laquelle constater les évolutions}

\item{sub}{sous-titre}

\item{horizontal}{agencement des images multiples: TRUE ou FALSE}

\item{overlay}{sf polygones à supperposer sur l'image}

\item{mask}{TRUE pour utiliser overlay comme masque}

\item{col}{couleur de délimitation du seuil}
}
\value{
plot
}
\description{
Test de délimitation des zones ouvertes entre 2 dates
}
