% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_plotDiff.R
\name{spot_plotDiff}
\alias{spot_plotDiff}
\title{Image de l'indice et de son évolution}
\usage{
spot_plotDiff(
  diff,
  pal1 = c("green", "yellow", "red"),
  pal2 = c("green", "green", "yellow", "red")
)
}
\arguments{
\item{diff}{liste issue de spot_differences}

\item{pal1}{vecteur de couleurs pour carte d'indice}

\item{pal2}{vcteur de couleurs pour carte des différence}
}
\value{
plot
}
\description{
Image de l'indice et de son évolution
}
