% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_plotRGB.R
\name{spot_plotRGB}
\alias{spot_plotRGB}
\title{Image couleur SPOT}
\usage{
spot_plotRGB(
  s,
  ir = FALSE,
  ans = 0,
  horizontal = TRUE,
  overlay = NULL,
  mask = FALSE,
  sub = NULL,
  evo = NULL,
  col = "blue"
)
}
\arguments{
\item{s}{spatraster 4 bandes nommées: r (rouge) ir (infrarouge) b (bleu) et g (vert), OU liste de spatraster 4 bandes nommées "spotXXXX" où XXXX est l'année de prise de vue}

\item{ir}{TRUE pour une image en fausses couleurs infrarouge}

\item{ans}{0 pour toutes les images, ou vecteur des années à afficher}

\item{horizontal}{agencement des images multiples: TRUE ou FALSE}

\item{overlay}{sf polygones à supperposer sur l'image}

\item{mask}{TRUE pour utiliser overlay comme masque}

\item{sub}{sous-titre}

\item{evo}{raster 0/1 des évolutions (facultatif)}

\item{col}{couleur de délimitation du seuil}
}
\value{
plot
}
\description{
Image couleur SPOT
}
