% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_saisie_morts.R
\name{spot_saisie_morts}
\alias{spot_saisie_morts}
\title{Saisie des points de calibration de mortalité}
\usage{
spot_saisie_morts(an1 = 2023, an0 = NULL)
}
\arguments{
\item{an1}{Année n+1}

\item{an0}{Année n (par défaut an1 - 1 an)}
}
\value{
table x, y, type (sec ou disparu)
}
\description{
Saisie des points de calibration de mortalité
}
