% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_select_year.R
\name{spot_select_year}
\alias{spot_select_year}
\title{Image SPOT d'une année donnée}
\usage{
spot_select_year(
  an,
  path_spot_ts = oiseauData::data_conf("path_spot_ts"),
  spot_best_day = dataOiseau::data_conf("spot_best_day"),
  ifnull = c("prev", "next", "nothing")
)
}
\arguments{
\item{an}{muneric}

\item{path_spot_ts}{chemin de la série temporelle spot de l'emprise du projet}

\item{spot_best_day}{MM-DD jour préférentiel}

\item{ifnull}{si non trouvée, chercher l'année précédente ("prev") ou suivante ("next") ou rien ("nothing")}
}
\value{
spatraster
}
\description{
Image SPOT d'une année donnée
}
