% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_spat2rast.R
\name{spot_spat2rast}
\alias{spot_spat2rast}
\title{Conversion SpatRaster en Raster}
\usage{
spot_spat2rast(spat)
}
\arguments{
\item{spat}{spatrasetr}
}
\value{
raster
}
\description{
Palie un bug de conversion de la projection
}
