% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_sun.R
\name{spot_sun}
\alias{spot_sun}
\title{Insolation à une date donnée}
\usage{
spot_sun(dem, month = 7, day = 21)
}
\arguments{
\item{dem}{raster modèle numérique d'élévation (MNT + MNH)}

\item{month}{mois}

\item{day}{jour}
}
\value{
raster d'insolation globale en J/m2
}
\description{
Insolation à une date donnée
}
