% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_untar.R
\name{spot_untar}
\alias{spot_untar}
\title{Import des images depuis le serveur THEIA}
\usage{
spot_untar(
  path_tar = NULL,
  dir = oiseauData::data_conf("dos_spot"),
  unlink_tar = FALSE
)
}
\arguments{
\item{path_tar}{chemin contenant les fichiers .tar.gz téléchargés}

\item{dir}{dossier dos_spot défini par oiseauData}

\item{unlink_tar}{TRUE pour supprimer les archives tar.gz}
}
\value{
décompresse selon l'architecture oiseau les fichiers présents
}
\description{
Voir la vignette GEOSUD
}
