% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_crowns_classify.R
\name{md_crowns_classify}
\alias{md_crowns_classify}
\title{Classification automatique des couronnes}
\usage{
md_crowns_classify(
  ncat = 8,
  rm_size = TRUE,
  dest_file = "class_crowns.tif",
  path_crowns = data_conf("path_crowns_ts"),
  path_spot = data_conf("path_spot_ts"),
  dest_dos = dc("dos_spot")
)
}
\arguments{
\item{ncat}{nb de types à créer}

\item{rm_size}{TRUE pour retirer les dimensions (h et surface) des facteurs explicatifs}

\item{dest_file}{nom du fichier à créer (avec extention .tif)}

\item{path_crowns}{chemin du raster des couronnes}

\item{path_spot}{chemin du rster des images spot}

\item{dest_dos}{dossier d'écriture}
}
\value{
spatRaster de catégories
}
\description{
Classification automatique des couronnes
}
