% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_model_ess_crowns.R
\name{md_model_ess_crowns}
\alias{md_model_ess_crowns}
\title{Création modèle prédiction des essences des couronnes}
\usage{
md_model_ess_crowns(path_cal, dest = NULL)
}
\arguments{
\item{path_cal}{chemin des couronnes de calibration créés avec md_ref_crowns_params()}

\item{dest}{NULL}
}
\value{

}
\description{
Création modèle prédiction des essences des couronnes
}
