% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_predict_crowns.R, R/md_predict_crowns_mlr3.R
\name{md_predict_crowns}
\alias{md_predict_crowns}
\title{Prédiction appliquée sur les couronnes}
\usage{
md_predict_crowns(
  rr,
  dir = dc("dos_modeles_oiseau"),
  path_predicteurs = file.path(dc("dos_projet"), "pile_essences.tif"),
  res = 1,
  spot_an = NULL,
  rsp_coper = FALSE,
  dest = dc("path_essences")
)

md_predict_crowns(
  rr,
  dir = dc("dos_modeles_oiseau"),
  path_predicteurs = file.path(dc("dos_projet"), "pile_essences.tif"),
  res = 1,
  spot_an = NULL,
  rsp_coper = FALSE,
  dest = dc("path_essences")
)
}
\arguments{
\item{rr}{objet de classe ResampleResult issu par ex de md_mlr3()}

\item{dir}{dossier du fichier}

\item{res}{resolution des rasters d'extraction (defaut tiré du nom du fichier ...-res1-... = 1m)}

\item{spot_an}{date de l'image spot uitilisée (defaut tiré du nom du fichier ...-spot2018-... = 2018)}

\item{rsp_coper}{TRUE pour rééchantilloner les rasters Coprenicus}

\item{dest}{}

\item{mn}{nom du modèle: nom du fichier où le modèle est enregistré, sans extension .rds}
}
\value{
spatraster

spatraster
}
\description{
Prédiction appliquée sur les couronnes

Prédiction appliquée sur les couronnes
}
