% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_ref_crowns_params.R
\name{md_ref_crowns_params}
\alias{md_ref_crowns_params}
\title{Métriques des houppiers des arbres des placettes permanentes}
\usage{
md_ref_crowns_params(
  nom_ref = "ess_ain",
  placettes = dbOiseau::bdloc_read_table("placettes_permanentes_placettes") \%>\%
    filter(!is.na(date) & date > as.Date("2015-01-01")),
  arbres_ttl = dbOiseau::bdloc_read_table("placettes_permanentes_arbres") \%>\%
    filter(is.na(type)),
  champ_essence = "groupe",
  champ_diam = "diam1",
  spec_date = 2018,
  topo = TRUE,
  dendro = TRUE,
  spot = TRUE,
  insol = TRUE,
  mnh = TRUE,
  sentinel = FALSE,
  copernicus = TRUE,
  rsp_coper = FALSE,
  dest_dir = dc("dos_modeles_oiseau"),
  user = basename(dirname(dc("dos_projet"))),
  test = FALSE,
  add = NULL
)
}
\arguments{
\item{nom_ref}{nom du jeu de données}

\item{placettes}{table des placettes permanentes, avec champ id = identifiant de la placette}

\item{arbres_ttl}{table des arbres avec champ de coordonnées polaire azimut (en grad) et dist (en m) et champs placette (id)}

\item{champ_essence}{de la table arbres}

\item{spec_date}{date des images spot à utiliser}

\item{topo}{TRUE pour utiliser}

\item{dendro}{TRUE pour utiliser}

\item{spot}{TRUE pour utiliser}

\item{insol}{TRUE pour utiliser}

\item{mnh}{TRUE pour utiliser}

\item{sentinel}{TRUE pour utiliser les données sentinel}

\item{copernicus}{TRUE pour utiliser les données copernicus}

\item{dest_dir}{dossier d'écriture}

\item{user}{nom utilisateur}

\item{test}{TRUE pour test sur 5 placettes}

\item{add}{NULL ou nom_ref du même jeu de crowns auquel ajouter les données extraites}

\item{champ_num}{(vecteur) des champs de la table arbre à conserver}
}
\value{
chemin du fichier d'écriture, écrit le fichier .rds des données de calibration
}
\description{
Métriques des houppiers des arbres des placettes permanentes
}
