% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_crowns.R
\name{spot_crowns}
\alias{spot_crowns}
\title{Détection des couronnes sur MNH}
\usage{
spot_crowns(
  path_mnh,
  ext,
  buffer,
  lim_h_rege = 3,
  pente = 0.07,
  intercept = 2,
  mask_mnh = NULL
)
}
\arguments{
\item{path_mnh}{chemin du MNH}

\item{ext}{polygone sf de l'étendue à détecter}

\item{buffer}{tampon à appliquer à exr}

\item{lim_h_rege}{hauteur minimale des apex à détecter}

\item{pente}{voir lidR::locate_trees}

\item{intercept}{voir lidR::locate_trees}

\item{mask_mnh}{spatraster 0/1, où 0 correspond aux couronnes disparues les années précédentes}
}
\value{
spatvector des couronnes
}
\description{
Détection des couronnes sur MNH
}
