% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_crowns.R
\name{spot_crowns}
\alias{spot_crowns}
\title{Détection des couronnes sur MNH}
\usage{
spot_crowns(
  date_mnh = "last",
  ext = oiseauData::data_conf("shp"),
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  dest_crowns_ts = oiseauData::data_conf("path_crowns_ts"),
  path_meta = oiseauData::data_conf("tab_crowns"),
  path_mnt = oiseauData::data_conf("path_mnt"),
  buffer = oiseauData::data_conf("buffer"),
  lim_h_rege = oiseauData::data_conf("lim_h_rege"),
  pente = 0.07,
  intercept = 2,
  smooth = 5,
  uniqueness = "incremental",
  return_apex = FALSE
)
}
\arguments{
\item{date_mnh}{du MNH utilisé}

\item{ext}{polygone sf de l'étendue à détecter}

\item{path_mnh_ts}{chemin du MNH}

\item{dest_crowns_ts}{chemin du fichier d'écriture}

\item{path_meta}{chemin du fichier des métadonnées}

\item{path_mnt}{chemin du MNT}

\item{buffer}{tampon à appliquer à exr}

\item{lim_h_rege}{hauteur minimale des apex à détecter}

\item{pente}{voir lidR::locate_trees}

\item{intercept}{voir lidR::locate_trees}

\item{smooth}{fenêtre de lissage du mnh}

\item{uniqueness}{incremental ou bitmerge}

\item{return_apex}{si TRUE, renvoie une liste: crowns et apex}
}
\value{
spatvector des couronnes ou liste si return_apex
}
\description{
Détection des couronnes sur MNH
}
