% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_data.R
\name{spot_data}
\alias{spot_data}
\title{Création image spot6/7 1.5m d'une zone}
\usage{
spot_data(
  dos_spot = oiseauData::data_conf("dos_spot"),
  roi = oiseauData::data_conf("shp"),
  destpath = oiseauData::data_conf("path_spot_ts"),
  buffer = oiseauData::data_conf("buffer"),
  info = FALSE,
  day_opti = "06-15"
)
}
\arguments{
\item{dos_spot}{dossier des images spot6/7}

\item{roi}{sf de la zone d'intérêt. Si NULL, renvoie seulement la liste des polygones des emprises}

\item{destpath}{chemin du dossier où écrire les images}

\item{buffer}{buffer à appliquer à roi}

\item{info}{TRUE pour ne renvoyer que le taux de couverture par image}

\item{day_opti}{jour de l'année optimal pour la prise de vue, format MM-JJ}
}
\value{
liste des polygones es emprises des dalles spots
}
\description{
Création image spot6/7 1.5m d'une zone
}
