% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_formula.R
\name{spot_formula}
\alias{spot_formula}
\title{Formules d'indices de végétation SPOT}
\usage{
spot_formula(
  indice = c("ndvi", "gli", "logr", "bai", "savi", "lai"),
  list = FALSE
)
}
\arguments{
\item{indice}{nom de l'indice. Voir https://www.indexdatabase.de/db/s-single.php?id=173}

\item{list}{TRUE pour renvoyer la liste des indices disponibles}
}
\value{
texte de la formule
}
\description{
Formules d'indices de végétation SPOT
}
