% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_indice.R
\name{spot_indice}
\alias{spot_indice}
\title{Indice de végétation des couronnes}
\usage{
spot_indice(
  date_crowns = "last",
  date_spot,
  crowns = TRUE,
  shp_roi = data_conf("shp"),
  path_spot_ts = data_conf("path_spot_ts"),
  path_crowns_ts = data_conf("path_crowns_ts"),
  path_tab_crowns = data_conf("tab_crowns"),
  path_mnt = data_conf("path_mnt"),
  buffer = data_conf("buffer"),
  indice = c("ndvi", "bai", "savi"),
  res = NULL
)
}
\arguments{
\item{date_crowns}{date du MNH utilisé, ou 'last' pour utiliser le plus récent}

\item{date_spot}{date de l'image SPOT utilisée}

\item{crowns}{TRUE pour l'indice par couronne, sinon par pixel}

\item{shp_roi}{sf de découpage}

\item{path_spot_ts}{chemin vers le raster de série temporelle spot}

\item{path_crowns_ts}{chemin vers le raster de série temporelle des couronnes (résolution 1m)}

\item{path_tab_crowns}{chemin du fichier meta de la série temporelle des couronnes}

\item{path_mnt}{chemin du MNT}

\item{buffer}{buffer à appliquer à shp_roi}

\item{indice}{noms des indices utilisés, par défaut ndvi. voir spot_formula}

\item{res}{NULL ou nouvelle résolution (m)}
}
\value{
spatRaster
}
\description{
Indice de végétation des couronnes
}
