% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_insolation.R
\name{spot_insolation}
\alias{spot_insolation}
\title{Raster d'insolation}
\usage{
spot_insolation(
  mnh = uRast("mnh"),
  mnt = uRast("mnt"),
  ext = dc("shp"),
  h_min = dc("lim_h_rege"),
  dest = dc("path_insolation_ts")
)
}
\arguments{
\item{mnh}{liste de SpatRaster de hauteur (séries temporelles)}

\item{mnt}{MNT}

\item{ext}{sf polygone de la zone à cartographier}

\item{h_min}{hauteur à partir du sol où évaluer le couvert}
}
\value{
liste de SpatRasetr d'insolation globale en J/m2
}
\description{
Raster d'insolation
}
