% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_maj_mnh.R
\name{spot_maj_mnh}
\alias{spot_maj_mnh}
\title{Mise à jour du MNH avec images spots}
\usage{
spot_maj_mnh(
  ext = data_conf("shp"),
  path_spot_ts = data_conf("path_spot_ts"),
  path_mnh_ts = data_conf("path_mnh_ts"),
  tab_mnh = data_conf("tab_mnh"),
  path_crowns_ts = data_conf("path_crowns_ts"),
  tab_crowns = data_conf("tab_crowns"),
  tab_deads = data_conf("tab_deads"),
  dest_deads = data_conf("path_deads_ts"),
  dest_mnh = data_conf("path_mnh_ts"),
  buffer = data_conf("buffer"),
  seuil_diff_spot = data_conf("seuil_diff_spot"),
  replace = FALSE
)
}
\arguments{
\item{ext}{sf de la zone}

\item{path_spot_ts}{chemin du fichier .nc d'écriture de la série temporelle}

\item{path_mnh_ts}{chemin du MNH}

\item{path_crowns_ts}{chemin vers le raster de série temporelle des couronnes (résolution 1m)}

\item{tab_crowns}{chemin du fichier meta de la série temporelle des couronnes}

\item{dest_deads}{chemin d'écriture du fichier raster .tif des masques}

\item{dest_mnh}{chemin d'écriture du fichier raster .tif des mnh}

\item{buffer}{buffer shp}

\item{seuil_diff_spot}{seuil de différence à partir duquel la couronne est éliminée}

\item{an_fin}{OBLOGATOIRE: dernière année à mettre à jour}

\item{.dir}{Dossier du projet si oiseauData est utilisé. Omet tous les autres arguments sauf an_fin}
}
\value{
liste des masques annuels à appliquer au MNT
}
\description{
L'utilisateur sélectionne 1 image par an puis les seuils de détection pour chaque année
}
