% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_maj_mnh.R
\name{spot_maj_mnh}
\alias{spot_maj_mnh}
\title{Mise à jour du MNH avec images spots}
\usage{
spot_maj_mnh(
  ext = oiseauData::data_conf("shp"),
  path_spot_ts = oiseauData::data_conf("path_spot_ts"),
  path_mnh_ts = oiseauData::data_conf("path_mnh_ts"),
  dest_masques = oiseauData::data_conf("path_mnh_dead_ts"),
  dest_mnh = oiseauData::data_conf("path_mnh_ts"),
  buffer = oiseauData::data_conf("buffer"),
  seuil_diff_spot = oiseauData::data_conf("seuil_diff_spot"),
  replace = FALSE
)
}
\arguments{
\item{ext}{sf de la zone}

\item{path_spot_ts}{chemin du fichier .nc d'écriture de la série temporelle}

\item{path_mnh_ts}{chemin du MNH}

\item{dest_masques}{chemin d'écriture du fichier raster .tif des masques}

\item{dest_mnh}{chemin d'écriture du fichier raster .tif des mnh}

\item{buffer}{buffer shp}

\item{seuil_diff_spot}{seuil de différence à partir duquel la couronne est éliminée}

\item{an_fin}{OBLOGATOIRE: dernière année à mettre à jour}

\item{.dir}{Dossier du projet si oiseauData est utilisé. Omet tous les autres arguments sauf an_fin}
}
\value{
liste des masques annuels à appliquer au MNT
}
\description{
L'utilisateur sélectionne 1 image par an puis les seuils de détection pour chaque année
}
