% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_maj_mnh.R
\name{spot_maj_mnh}
\alias{spot_maj_mnh}
\title{Mise à jour du MNH avec images spots}
\usage{
spot_maj_mnh(ext, dos_spot, path_data, path_mnh, an_mnh, an_fin)
}
\arguments{
\item{ext}{sf de la zone}

\item{dos_spot}{dossier des images spot}

\item{path_data}{dossier où écrire les résultats}

\item{path_mnh}{chemin du MNH}

\item{an_mnh}{année du MNH}

\item{an_fin}{dernière année à mettre à jour}
}
\value{
liste des masques annuels à appliquer au MNT
}
\description{
L'utilisateur sélectionne 1 image par an puis les seuils de détection pour chaque année
}
