% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_plotDiff.R
\name{spot_plotDiff}
\alias{spot_plotDiff}
\title{Image de l'indice et de son évolution}
\usage{
spot_plotDiff(
  diff,
  ID = NULL,
  overlay = data_conf("shp"),
  pal1 = c("green", "yellow", "red"),
  pal2 = c("green", "green", "yellow", "red"),
  rev = TRUE
)
}
\arguments{
\item{diff}{liste issue de spot_differences}

\item{ID}{zoom sur un ID foncier ou NULL pour tout}

\item{overlay}{couche vecteur à superposer (à défaut le shp foncier)}

\item{pal1}{vecteur de couleurs pour carte d'indice}

\item{pal2}{vcteur de couleurs pour carte des différence}

\item{rev}{TRUE si la disparition correspond à une baisse de l'indice, FALSE sinon}
}
\value{
plot
}
\description{
Image de l'indice et de son évolution
}
